+++ lib/cldr/calendar/backend/calendar.ex
@@ -28,6 +28,81 @@ defmodule Cldr.Calendar.Backend do
 
         @default_calendar :gregorian
 
+        @doc """
+        Returns a keyword list of options than can be applied to
+        `NimbleStrftime.format/3`.
+
+        The hex package [nimble_strftime](https://hex.pm/packages/nimble_strftime)
+        provides a `format/3` function to format dates, times and datetimes.
+        It takes a set of options that can return day, month and am/pm names.
+
+        `strftime_options!` returns a keyword list than can be used as these
+        options to return localised names for days, months and am/pm.
+
+        ## Arguments
+
+        * `locale` is any locale returned by `MyApp.Cldr.known_locales/0`. The
+          default is `MyApp.Cldr.get_locale/0`
+
+        * `options` is a set of keyword options. The default is `[]`
+
+        ## Options
+
+        * `:calendar` is the name of any known CLDR calendar. The default
+          is `:gregorian`.
+
+        ## Example
+
+            iex: MyApp.Cldr.Calendar.strftime_options!
+            [
+              am_pm_names: #Function<0.32021692/1 in MyApp.Cldr.Calendar.strftime_options/2>,
+              month_names: #Function<1.32021692/1 in MyApp.Cldr.Calendar.strftime_options/2>,
+              abbreviated_month_names: #Function<2.32021692/1 in MyApp.Cldr.Calendar.strftime_options/2>,
+              day_of_week_names: #Function<3.32021692/1 in MyApp.Cldr.Calendar.strftime_options/2>,
+              abbreviated_day_of_week_names: #Function<4.32021692/1 in MyApp.Cldr.Calendar.strftime_options/2>
+            ]
+
+        ## Typical usage
+
+            iex: NimbleStrftime.format(Date.today(), MyApp.Cldr.Calendar.strftime_options!())
+
+        """
+
+        def strftime_options!(locale \\ unquote(backend).get_locale(), options \\ []) do
+          calendar = Keyword.get(options, :calendar, @default_calendar)
+
+          with {:ok, locale} <- Cldr.validate_locale(locale) do
+            [
+              am_pm_names: fn am_pm ->
+                day_periods(locale, calendar)
+                |> get_in([:format, :abbreviated, am_pm])
+              end,
+
+              month_names: fn month ->
+                months(locale, calendar)
+                |> get_in([:format, :wide, month])
+              end,
+
+              abbreviated_month_names: fn month ->
+                months(locale, calendar)
+                |> get_in([:format, :abbreviated, month])
+              end,
+
+              day_of_week_names: fn day ->
+                days(locale, calendar)
+                |> get_in([:format, :wide, day])
+              end,
+
+              abbreviated_day_of_week_names: fn day ->
+                days(locale, calendar)
+                |> get_in([:format, :abbreviated, day])
+              end
+            ]
+          else
+            {:error, {exception, message}} -> raise exception, message
+          end
+        end
+
         def eras(locale \\ unquote(backend).get_locale(), calendar \\ @default_calendar)
 
         def eras(%LanguageTag{cldr_locale_name: cldr_locale_name}, calendar) do
@@ -108,6 +183,7 @@ defmodule Cldr.Calendar.Backend do
 
           def day_periods(unquote(locale_name), calendar),
             do: {:error, Calendar.calendar_error(calendar)}
+
         end
 
         def eras(locale, _calendar), do: {:error, Locale.locale_error(locale)}